python-backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
langgraph>=0.2.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langsmith>=0.1.0
openai>=1.0.0
anthropic>=0.7.0
pydantic>=2.0.0
asyncio-mqtt>=0.11.0
redis>=5.0.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0
httpx>=0.25.0
aiofiles>=23.0.0
structlog>=23.0.0
tenacity>=8.0.0
